Revert "Merge pull request #10 from danidiaaz27/fix/Pos"
This reverts commit 172b43a1530ce81bea79bf5d460628a317bb636d, reversing
changes made to 22f97b2ef6e4f040028d1caa91438fd6c26182b7.

+++ sistema/src/main/java/com/hatunvet/sistema/controller/InventarioController.java
@@ -81,7 +81,7 @@ public class InventarioController {
             Producto producto = prodOpt.get();
             movimiento.setProducto(producto);
 
-            double stockActual = producto.getStock();
+            int stockActual = producto.getStock();
             int cantidadMovimiento = movimiento.getCantidad(); // Ya sabemos que es positivo seguro
 
             // LOGICA MATEMATICA SEGURA
